GH-83 Move eligible party demo application into an examples folder and rename it to example-app
