+++ examples/build.gradle.kts
@@ -1,0 +1,19 @@
+plugins {
+    id("energy.eddie.java-conventions")
+}
+
+group = "energy.eddie"
+version = "0.0.0"
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    testImplementation("org.junit.jupiter:junit-jupiter-api:5.8.1")
+    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.8.1")
+}
+
+tasks.getByName<Test>("test") {
+    useJUnitPlatform()
+}
