+++ examples/example-app/src/main/docker/Dockerfile
@@ -1,5 +1,5 @@
 FROM eclipse-temurin:20-jre-alpine
 COPY ./build/install/ /opt/
-WORKDIR /opt/ep-demo-app
-CMD ["sh", "./bin/ep-demo-app"]
+WORKDIR /opt/example-app
+CMD ["sh", "./bin/example-app"]
 EXPOSE 8080
