+++ examples/example-app/src/main/java/energy/eddie/examples/exampleapp/ConnectionStatus.java
@@ -1,4 +1,4 @@
-package energy.eddie.epdemoapp;
+package energy.eddie.examples.exampleapp;
 
 public record ConnectionStatus(String connectionId, Object timestamp, String status) {
 }
