+++ examples/example-app/src/main/java/energy/eddie/examples/exampleapp/Env.java
@@ -1,4 +1,4 @@
-package energy.eddie.epdemoapp;
+package energy.eddie.examples.exampleapp;
 
 public enum Env {
     JDBC_URL, JDBC_USER, JDBC_PASSWORD, EDDIE_FRAMEWORK_PUBLIC_URL, PUBLIC_CONTEXT_PATH;
