+++ examples/example-app/src/main/java/energy/eddie/examples/exampleapp/JavalinHandler.java
@@ -1,4 +1,4 @@
-package energy.eddie.epdemoapp;
+package energy.eddie.examples.exampleapp;
 
 import io.javalin.Javalin;
