+++ examples/example-app/src/main/jte/show-connection.jte
